Move protooptions to sdk/pbs
This prevents a reverse dependency from sdk on the main module.
